fix(wallet): address remaining review comments on Mine Blocks dialog
- Change MineBlocksSuccess(usize) to MineBlocksSuccess(u64) for
  type consistency with block_count parameter (Claudius #5)
- Align dialog close pattern with Send/Receive: use local `open`
  variable for egui X button, reset state inside closure for
  Cancel/Mine buttons (Claudius #6)

+++ src/backend_task/core/mod.rs
@@ -300,7 +300,7 @@ impl AppContext {
                 .await
                 .map_err(|e| format!("Task join error: {}", e))??;
 
-                let mined_count = mined.len();
+                let mined_count = mined.len() as u64;
 
                 // Refresh wallet balances via RPC so the UI reflects the new coins
                 let refresh_ctx = self.clone();
